(fix) add previous button to UI
I removed the Profile Questions controller and template, because it
actually has the same functionality as find a date. I also turned the
tags from an array into an object, so that if the user changes their
mind about things in the UI, they can overwrite their previous choices.

I moved the dateGenre questions into the controllers file, too, and
wrote some functions (like submitSoFar and preloadChosenOption) to help
handle when the user clicks "Next" and "Previous".

PLEASE NOTE, the way that this currently stands, the user can only
choose one "tag" on the first screen, as opposed to multiple ones
before.